requirements-prod.txt
# Production requirements - minimal set for running the API
# Excludes development, testing, and type checking packages

# Core Framework
aiofiles==24.1.0
fastapi==0.115.14
uvicorn[standard]==0.35.0
websockets==15.0.1

# Pydantic
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-ai==0.3.5

# Database and Storage
supabase==2.16.0
redis==6.2.0
asyncpg==0.30.0
pgvector==0.4.1

# Authentication and Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
slowapi==0.1.9

# AI/ML - Core only
openai==1.93.0
anthropic==0.57.1

# Essential utilities
python-dotenv==1.1.1
httpx==0.28.1
anyio==4.9.0

# Logging (optional but recommended)
logfire==3.22.0

# Required dependencies that might be missing
starlette==0.46.2
typing-extensions==4.14.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
requests==2.32.4
sniffio==1.3.1
click==8.2.1
h11==0.16.0
httpcore==1.0.9
pycparser==2.22
cffi==1.17.1
cryptography==45.0.5
ecdsa==0.19.1
pyasn1==0.6.1
rsa==4.9.1
six==1.17.0
bcrypt==4.3.0
PyJWT==2.10.1

# File validation dependencies
python-magic==0.4.27
python-magic-bin==0.4.14
Pillow==11.3.0